requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.0.1
python-multipart==0.0.6
easyocr==1.7.2
fitz
PyMuPDF